refactor: changed chat info to channel in mongoldb
